WordPress Plugin Updater Script
This commit adds a Python script that updates all instances of a plugin installed on a similar machine. The script prompts the user to enter the name of the plugin they wish to update, and then searches the machine for all instances of that plugin. It then downloads the specified version of the plugin from the official repository and installs it in all instances found.